requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-cors-headers==4.8.0
python-decouple==3.8
google-generativeai==0.8.3
twilio==9.2.4
requests==2.32.3
Pillow==10.4.0
django-filter==24.3